WIP: ASM gen handles rf-arity-2-arg-1 type

+++ src-java/xfseq/buffer/DoubleBuffer.java
@@ -28,7 +28,7 @@ public class DoubleBuffer extends AFn implements IXFSeqBuffer, IFn.ODO {
     }
 
     public Object invoke(Object a, Object b) {
-        return invokePrim(a, (Long)b);
+        return invokePrim(a, (Double)b);
     }
 
     @Override
